fix: modalDeleteCategories on listCategories
