+++ src/pages/ListCategory/ListCategory.jsx
@@ -6,7 +6,10 @@ import { TbPencil } from 'react-icons/tb';
 import { useMediaQuery } from 'react-responsive';
 import { toast } from 'react-toastify';
 
-import { ModalEditCategory } from '../../components/features';
+import {
+  ModalDeleteCategory,
+  ModalEditCategory,
+} from '../../components/features';
 import { useSearchByNameCategories } from '../../hooks/query/categories';
 import useDebounce from '../../hooks/useDebounce';
 import {
@@ -22,6 +25,7 @@ import {
   SearchIconButton,
   SearchSection,
   ModalStyle,
+  DeleteButton,
 } from './Styles';
 import { buildGetCategoriesErrorMessage } from './utils';
 
@@ -29,6 +33,8 @@ export default function ListCategory() {
   const isSmallScreen = useMediaQuery({ maxWidth: 700 });
   const [modalEditCategory, setModalEditCategory] = useState(false);
   const [selectedCategory, setSelectedCategory] = useState({});
+  const [modalDeleteCategory, setModalDeleteCategory] = useState(false);
+  const [categoryId, setCategoryId] = useState('');
   const [name, setName] = useState('');
   const debouncedName = useDebounce(name);
 
@@ -45,6 +51,12 @@ export default function ListCategory() {
     setModalEditCategory(true);
   };
   const closeModalEditCategory = () => setModalEditCategory(false);
+
+  const openModalDeleteCategory = (_id) => {
+    setCategoryId(_id);
+    setModalDeleteCategory(true);
+  };
+  const closeModalDeleteCategory = () => setModalDeleteCategory(false);
 
   const modalCloseButton = <CloseOutlined style={{ color: 'white' }} />;
   return (
@@ -67,6 +79,11 @@ export default function ListCategory() {
         {categories?.map((category) => (
           <Row key={category._id}>
             <Text>{category.name}</Text>
+            <DeleteButton>
+              <CloseOutlined
+                onClick={() => openModalDeleteCategory(category._id)}
+              />
+            </DeleteButton>
 
             {isSmallScreen ? (
               <StyledLink to="/administrador/editar-categoria" state={category}>
@@ -106,6 +123,24 @@ export default function ListCategory() {
         <ModalEditCategory
           category={selectedCategory}
           close={closeModalEditCategory}
+        />
+      </ModalStyle>
+      <ModalStyle
+        open={modalDeleteCategory}
+        onCancel={closeModalDeleteCategory}
+        footer={null}
+        width={1000}
+        closeIcon={<CloseOutlined />}
+        destroyOnClose
+        centered
+        bodyStyle={{
+          background: '#123645',
+          color: 'white',
+        }}
+      >
+        <ModalDeleteCategory
+          _id={categoryId}
+          close={closeModalDeleteCategory}
         />
       </ModalStyle>
     </Container>
