+++ src/pages/ListProduct/Styles.js
@@ -118,8 +118,9 @@ export const SearchProduct = styled.input`
 
 export const Row = styled.div`
   display: grid;
-  grid-template-columns: 30px 1fr 1fr 30px 30px;
+  grid-template-columns: 1fr 1fr 30px 30px;
   justify-content: space-between;
+  align-items: center;
   gap: 10px;
   width: 100%;
 `;
